requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
Pillow==10.2.0
reportlab==4.0.9
firebase-admin==6.4.0
python-decouple==3.8
PyPDF2==3.0.1
